requirements.txt
python-dateutil>=2.8.2
requests>=2.31.0
icalendar>=5.0.11
python-dotenv>=1.0.0
ollama>=0.1.6
mcp-client>=0.1.0
flask>=2.0.0
flask-cors>=3.0.10 